Datagrid pages in northwind seems functional

+++ src/pages/datagridInternals/styledSimpleDatagrid.ts
@@ -5,17 +5,8 @@ import { insertDialog, InsertDialogOpts } from "../../components/insertDialog";
 import { materialIcon } from "../../components/materialIcon";
 import { typography } from "../../components/typography";
 import { Table } from "../../app";
-import { element, ifNode, sourceMap, state } from "../../nodeHelpers";
+import { nodes } from "../../nodeHelpers";
 import { Node } from "../../nodeTypes";
-import {
-  commitTransaction,
-  if_,
-  modify,
-  scalar,
-  serviceProc,
-  setScalar,
-  startTransaction,
-} from "../../procHelpers";
 import { createStyles, flexGrowStyles } from "../../styleUtils";
 import { ident, stringLiteral } from "../../utils/sqlHelpers";
 import { SqlExpression, StateStatement } from "../../yom";
@@ -29,6 +20,7 @@ import { styles as sharedStyles } from "./styles";
 import { triggerQueryRefresh } from "./shared";
 import { Cell, ColumnEventHandlers, RowHeight } from "./types";
 import { circularProgress } from "../../components/circularProgress";
+import { StateStatements } from "../../statements";
 
 export interface ToolbarConfig {
   header: Node;
@@ -53,7 +45,7 @@ export interface StyledSimpleGridConfig {
   columns: SimpleColumn[];
   idField: string;
   pageSize?: number;
-  extraState?: StateStatement[];
+  extraState?: StateStatements;
   allow?: SqlExpression;
   sourceMapName?: string;
   rowHeight?: RowHeight;
@@ -94,31 +86,30 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
   } else if (config.toolbar.add?.type === "dialog") {
     const withValues: Record<string, string> =
       config.toolbar.add.opts?.withValues ?? {};
-    addButton = state({
-      procedure: [scalar(`adding`, `false`)],
+    addButton = nodes.state({
+      procedure: (s) => s.scalar(`adding`, `false`),
       children: [
         iconButton({
           variant: "soft",
           color: "primary",
           size: "sm",
           children: materialIcon("Add"),
-          on: { click: [setScalar(`adding`, `true`)] },
+          on: { click: (s) => s.setScalar(`adding`, `true`) },
         }),
         insertDialog({
           ...config.toolbar.add.opts,
           table: config.tableModel.name,
           open: `adding`,
-          onClose: [setScalar(`adding`, `false`)],
+          onClose: (s) => s.setScalar(`adding`, `false`),
           content: {
             type: "AutoLabelOnLeft",
             ignoreFields: Object.keys(withValues),
           },
-          afterTransactionCommit: (state) => [
-            ...((config.toolbar.add as any).opts?.afterTransactionCommit?.(
-              state
-            ) ?? []),
-            triggerQueryRefresh(),
-          ],
+          afterTransactionCommit: (state, s) =>
+            s.statements(
+              (config.toolbar.add as any).opts?.afterTransactionCommit?.(state),
+              triggerQueryRefresh()
+            ),
         }),
       ],
     });
@@ -127,21 +118,21 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
     source: "db." + ident(config.tableModel.name),
     idFieldSource: ident(config.tableModel.primaryKeyFieldName),
     children: (dg) =>
-      ifNode(
-        `status = 'failed'`,
-        alert({
+      nodes.if({
+        expr: `status = 'failed'`,
+        then: alert({
           color: "danger",
           children: `'Unable to load data.'`,
         }),
-        [
-          element("div", {
+        else: [
+          nodes.element("div", {
             styles: styles.toolbar,
             children: [
               typography({
                 level: "h4",
                 children: config.toolbar.header,
               }),
-              ifNode(
+              nodes.if(
                 `status = 'fallback_triggered' and dg_refresh_key != 0`,
                 typography({
                   startDecorator: circularProgress({ size: "sm" }),
@@ -149,7 +140,7 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
                   children: `'Reloading...'`,
                 })
               ),
-              ifNode(
+              nodes.if(
                 `saving_edit`,
                 typography({
                   startDecorator: circularProgress({ size: "sm" }),
@@ -157,7 +148,7 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
                   children: `'Saving change...'`,
                 })
               ),
-              ifNode(
+              nodes.if(
                 `display_error_message is not null`,
                 alert({
                   startDecorator: materialIcon("Report"),
@@ -167,7 +158,7 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
                   children: `display_error_message`,
                 })
               ),
-              ifNode(
+              nodes.if(
                 `status = 'failed'`,
                 alert({
                   startDecorator: materialIcon("Report"),
@@ -177,29 +168,29 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
                   children: `'Failed to load data'`,
                 })
               ),
-              element("div", { styles: flexGrowStyles }),
+              nodes.element("div", { styles: flexGrowStyles }),
               config.toolbar.delete
-                ? state({
-                    procedure: [scalar(`deleting`, `false`)],
+                ? nodes.state({
+                    procedure: (s) => s.scalar(`deleting`, `false`),
                     children: [
                       iconButton({
                         size: "sm",
                         color: "danger",
                         variant: "soft",
                         children: materialIcon("DeleteOutlined"),
-                        on: { click: [setScalar(`deleting`, `true`)] },
+                        on: { click: (s) => s.setScalar(`deleting`, `true`) },
                       }),
                       confirmDangerDialog({
                         open: `deleting`,
-                        onClose: [setScalar(`deleting`, `false`)],
-                        description: element("span", {
+                        onClose: (s) => s.setScalar(`deleting`, `false`),
+                        description: nodes.element("span", {
                           children: [
                             `'Are you sure you want to delete '`,
-                            ifNode(
-                              `selected_all`,
-                              state({
-                                procedure: [
-                                  scalar(
+                            nodes.if({
+                              expr: `selected_all`,
+                              then: nodes.state({
+                                procedure: (s) =>
+                                  s.scalar(
                                     `count`,
                                     "(" +
                                       getCountQuery(
@@ -207,39 +198,37 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
                                       ) +
                                       ")"
                                   ),
-                                ],
                                 children: `count`,
                               }),
-                              `(select count(*) from selected_row)`
-                            ),
+                              else: `(select count(*) from selected_row)`,
+                            }),
                             `' records?'`,
                           ],
                         }),
-                        onConfirm: (closeModal) => [
-                          serviceProc([
-                            startTransaction(),
-                            if_(
-                              `selected_all`,
-                              [
-                                modify(
-                                  `delete from db.${ident(
-                                    config.tableModel.name
-                                  )}`
-                                ),
-                              ],
-                              [
-                                modify(
-                                  `delete from db.${ident(
-                                    config.tableModel.name
-                                  )} where id in (select id from ui.selected_row)`
-                                ),
-                              ]
-                            ),
-                            commitTransaction(),
-                            triggerQueryRefresh(),
-                          ]),
-                          ...closeModal,
-                        ],
+                        onConfirm: (closeModal) => (s) =>
+                          s
+                            .serviceProc((s) =>
+                              s
+                                .startTransaction()
+                                .if({
+                                  condition: `selected_all`,
+                                  then: (s) =>
+                                    s.modify(
+                                      `delete from db.${ident(
+                                        config.tableModel.name
+                                      )}`
+                                    ),
+                                  else: (s) =>
+                                    s.modify(
+                                      `delete from db.${ident(
+                                        config.tableModel.name
+                                      )} where id in (select id from ui.selected_row)`
+                                    ),
+                                })
+                                .commitTransaction()
+                                .statements(triggerQueryRefresh())
+                            )
+                            .statements(closeModal),
                       }),
                     ],
                   })
@@ -247,15 +236,15 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
               addButton,
             ],
           }),
-          ifNode(
-            `(status = 'requested' or status = 'fallback_triggered') and dg_refresh_key = 0`,
-            element("div", {
+          nodes.if({
+            expr: `(status = 'requested' or status = 'fallback_triggered') and dg_refresh_key = 0`,
+            then: nodes.element("div", {
               styles: sharedStyles.emptyGrid,
             }),
-            dg
-          ),
-        ]
-      ),
+            else: dg,
+          }),
+        ],
+      }),
     columns: baseColumns,
     idField: config.idField,
     datagridStyles: {
@@ -271,7 +260,7 @@ export function styledSimpleDatagrid(config: StyledSimpleGridConfig) {
     rowHeight: config.rowHeight ?? "medium",
   });
   if (config.sourceMapName) {
-    content = sourceMap(config.sourceMapName, content);
+    content = nodes.sourceMap(config.sourceMapName, content);
   }
   return content;
 }
